Export more toolkit-related things
