+++ src/index.js
@@ -47,6 +47,7 @@ export { default as ChatHistory } from './components/chat-history.jsx';
 export { default as AskUserComponent } from './components/ask-user.jsx';
 export { default as ConfirmComponent } from './components/confirm.jsx';
 export { default as MessageContent } from './components/message-content.jsx';
+export { default as MessageInput } from './components/message-input.jsx';
 export { default as UserMessageInput } from './components/user-message-input.jsx';
 
 /**
